// src/utils/readme-header-type.ts
/**
 * Readme header's type.
 */
enum ReadmeHeaderType {
    /**
     * Represents the latest header design.
     */
    New = "new",

    /**
     * Represents the latest header design, but with a book octicon in use.
     */
    CombinedNew = "combined-new",

    /**
     * Represents an old school design, but with the new functionality.
     */
    CombinedOld = "combined-old",

    /**
     * Represents an old school design.
     */
    Old = "old",
}

export default ReadmeHeaderType;